refactor showTx()
