+++ app/src/main/java/com/samourai/wallet/SendActivity.java
@@ -2129,165 +2129,4 @@ public class SendActivity extends Activity {
 
     }
 
-    private void doShowTx(final String hexTx, final String txHash) {
-
-        final int QR_ALPHANUM_CHAR_LIMIT = 4296;    // tx max size in bytes == 2148
-
-        TextView showTx = new TextView(SendActivity.this);
-        showTx.setText(hexTx);
-        showTx.setTextIsSelectable(true);
-        showTx.setPadding(40, 10, 40, 10);
-        showTx.setTextSize(18.0f);
-
-        final CheckBox cbMarkInputsUnspent = new CheckBox(SendActivity.this);
-        cbMarkInputsUnspent.setText(R.string.mark_inputs_as_unspendable);
-        cbMarkInputsUnspent.setChecked(false);
-
-        LinearLayout hexLayout = new LinearLayout(SendActivity.this);
-        hexLayout.setOrientation(LinearLayout.VERTICAL);
-        hexLayout.addView(cbMarkInputsUnspent);
-        hexLayout.addView(showTx);
-
-        new AlertDialog.Builder(SendActivity.this)
-                .setTitle(txHash)
-                .setView(hexLayout)
-                .setCancelable(false)
-                .setPositiveButton(R.string.close, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                        if(cbMarkInputsUnspent.isChecked())    {
-                            UTXOFactory.getInstance(SendActivity.this).markUTXOAsUnspendable(hexTx);
-                            Intent intent = new Intent("com.samourai.wallet.BalanceFragment.REFRESH");
-                            intent.putExtra("notifTx", false);
-                            intent.putExtra("fetch", true);
-                            LocalBroadcastManager.getInstance(SendActivity.this).sendBroadcast(intent);
-                        }
-
-                        dialog.dismiss();
-                        SendActivity.this.finish();
-
-                    }
-                })
-                .setNeutralButton(R.string.copy_to_clipboard, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                        android.content.ClipboardManager clipboard = (android.content.ClipboardManager)SendActivity.this.getSystemService(android.content.Context.CLIPBOARD_SERVICE);
-                        android.content.ClipData clip = null;
-                        clip = android.content.ClipData.newPlainText("TX", hexTx);
-                        clipboard.setPrimaryClip(clip);
-                        Toast.makeText(SendActivity.this, R.string.copied_to_clipboard, Toast.LENGTH_SHORT).show();
-
-                    }
-                })
-                .setNegativeButton(R.string.show_qr, new DialogInterface.OnClickListener() {
-                    public void onClick(DialogInterface dialog, int whichButton) {
-
-                        if(cbMarkInputsUnspent.isChecked())    {
-                            UTXOFactory.getInstance(SendActivity.this).markUTXOAsUnspendable(hexTx);
-                            Intent intent = new Intent("com.samourai.wallet.BalanceFragment.REFRESH");
-                            intent.putExtra("notifTx", false);
-                            intent.putExtra("fetch", true);
-                            LocalBroadcastManager.getInstance(SendActivity.this).sendBroadcast(intent);
-                        }
-
-                        if(hexTx.length() <= QR_ALPHANUM_CHAR_LIMIT)    {
-
-                            final ImageView ivQR = new ImageView(SendActivity.this);
-
-                            Display display = (SendActivity.this).getWindowManager().getDefaultDisplay();
-                            Point size = new Point();
-                            display.getSize(size);
-                            int imgWidth = Math.max(size.x - 240, 150);
-
-                            Bitmap bitmap = null;
-
-                            QRCodeEncoder qrCodeEncoder = new QRCodeEncoder(hexTx, null, Contents.Type.TEXT, BarcodeFormat.QR_CODE.toString(), imgWidth);
-
-                            try {
-                                bitmap = qrCodeEncoder.encodeAsBitmap();
-                            } catch (WriterException e) {
-                                e.printStackTrace();
-                            }
-
-                            ivQR.setImageBitmap(bitmap);
-
-                            LinearLayout qrLayout = new LinearLayout(SendActivity.this);
-                            qrLayout.setOrientation(LinearLayout.VERTICAL);
-                            qrLayout.addView(ivQR);
-
-                            new AlertDialog.Builder(SendActivity.this)
-                                    .setTitle(txHash)
-                                    .setView(qrLayout)
-                                    .setCancelable(false)
-                                    .setPositiveButton(R.string.close, new DialogInterface.OnClickListener() {
-                                        public void onClick(DialogInterface dialog, int whichButton) {
-
-                                            dialog.dismiss();
-                                            SendActivity.this.finish();
-
-                                        }
-                                    })
-                                    .setNegativeButton(R.string.share_qr, new DialogInterface.OnClickListener() {
-                                        public void onClick(DialogInterface dialog, int whichButton) {
-
-                                            String strFileName = AppUtil.getInstance(SendActivity.this).getReceiveQRFilename();
-                                            File file = new File(strFileName);
-                                            if(!file.exists()) {
-                                                try {
-                                                    file.createNewFile();
-                                                }
-                                                catch(Exception e) {
-                                                    Toast.makeText(SendActivity.this, e.getMessage(), Toast.LENGTH_SHORT).show();
-                                                }
-                                            }
-                                            file.setReadable(true, false);
-
-                                            FileOutputStream fos = null;
-                                            try {
-                                                fos = new FileOutputStream(file);
-                                            }
-                                            catch(FileNotFoundException fnfe) {
-                                                ;
-                                            }
-
-                                            if(file != null && fos != null) {
-                                                Bitmap bitmap = ((BitmapDrawable)ivQR.getDrawable()).getBitmap();
-                                                bitmap.compress(Bitmap.CompressFormat.PNG, 0, fos);
-
-                                                try {
-                                                    fos.close();
-                                                }
-                                                catch(IOException ioe) {
-                                                    ;
-                                                }
-
-                                                Intent intent = new Intent();
-                                                intent.setAction(Intent.ACTION_SEND);
-                                                intent.setType("image/png");
-                                                if (android.os.Build.VERSION.SDK_INT >= 24) {
-                                                    //From API 24 sending FIle on intent ,require custom file provider
-                                                    intent.putExtra(Intent.EXTRA_STREAM, FileProvider.getUriForFile(
-                                                            SendActivity.this,
-                                                            getApplicationContext()
-                                                                    .getPackageName() + ".provider", file));
-                                                } else {
-                                                    intent.putExtra(Intent.EXTRA_STREAM, Uri.fromFile(file));
-                                                }
-                                                startActivity(Intent.createChooser(intent, SendActivity.this.getText(R.string.send_tx)));
-                                            }
-
-                                        }
-                                    }).show();
-                        }
-                        else    {
-
-                            Toast.makeText(SendActivity.this, R.string.tx_too_large_qr, Toast.LENGTH_SHORT).show();
-
-                        }
-
-                    }
-                }).show();
-
-    }
-
 }
